Addressing imposter syndrome
I believe that, as part of the lab's discussion of how to address the impacts of systemic racism in the United States in an academic context, it may also be helpful to offer guidance to mentors on how to support their mentees who may be more likely to experience imposter syndrome in addition to other challenges. If this is outside the scope of the lab agreement and would be more valuable as a training exercise, or if it refocuses the conversation in an unhelpful direction, I also understand that decision. Here's just a starting point! Thanks for your time